Contributes to #1297.

@krizhanovsky said very well:

len in ss_skb_actor_t() comes from skb_headlen(), which is unsigned int. unsigned int headlen also makes the code more robust.

What was done here:
 - used unsigned int instead of size_t in ss_skb_actor_t;
 - also unsigned int is used instead of size_t at major network data processing points.

+++ fw/http_frame.c
@@ -1481,7 +1481,7 @@ conn_term:
  * Main FSM for processing HTTP/2 frames.
  */
 static int
-tfw_h2_frame_recv(void *data, unsigned char *buf, size_t len,
+tfw_h2_frame_recv(void *data, unsigned char *buf, unsigned int len,
 		  unsigned int *read)
 {
 	int n, r = T_POSTPONE;
